+++ apps/commerce-api/src/main/java/com/loopers/application/payment/processor/PaymentProcessor.java
@@ -1,0 +1,7 @@
+package com.loopers.application.payment.processor;
+
+import com.loopers.application.payment.PaymentResult;
+
+public interface PaymentProcessor {
+    PaymentResult process(PaymentContext context);
+}
