+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/payment/PaymentV1Dto.java
@@ -20,7 +20,7 @@ public class PaymentV1Dto {
                     info.orderId(),
                     info.paymentId(),
                     info.ourStatus(),
-                    info.pgStatus(),
+                    info.pgStatus().name(),
                     info.isSync(),
                     info.message()
             );
